[aon_timer] Workaround for concurrent resets
This is not really the right fix, but extends the previous fix
so that the resets are made concurrent (even by routines that call
apply_reset and not apply_reset_concurrently).

The right fix probably needs to happen in the dv_base somewhere so that
there is a notion of what kind of reset relationship we want.

Leaving that to another day and just get the regressions cleaner for now.
This resolves both csr_outstanding and stress_all failures.

Signed-off-by: Timothy Chen <timothytim@google.com>