[dv/kmac] Fix regression error
Issue #16681 reports an assertion error.
The root cause for this error is the entropy wait timer expired.
However, in normal cases, the testbench does not intent to trigger the
timer expired error. This is because when calculating the min wait time,
I did not count for each KMAC requests four entropies.

So this assertion error should be resolved if the timer expired
threshold sets to a larger value.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>