[opentitantool] Convenience methods for HyperDebug transport
The existing `execute_command()` invokes a callback for each line of
response received.  This would be an efficient way of handling megabytes
of data, without having to keep it all in memory at the same time.
However, most actual use cases expect a single line of output, and it is
cumbersome for each caller to have to count callbacks, and handle the
error cases of receving none or more than one.

This PR adds convenience methods to transport/hyperdebug for doing
textual request/response, when expecting no output, a single line of
output, or even a single line of output which matches a given regular
expression.  This way, the error cases can be centrally dealt with.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I18979e53265c6a8c85e6a77d5b141c704ebc4011